actualizacion de php exceptions

+++ server/deleteComp.php
@@ -10,7 +10,18 @@ if (empty($_GET["idComp"])) {
     exit("No hay id de componente para eliminar");
 }
 $idComp = $_GET["idComp"];
-$bd = include_once "bd.php";
-$sentencia = $bd->prepare("DELETE FROM components WHERE component_id = ?");
-$resultado = $sentencia->execute([$idComp]);
-echo json_encode($resultado);
+if (is_numeric($idComp)) {
+    try {
+        $bd = include_once "bd.php";
+        $sentencia = $bd->prepare("DELETE FROM components WHERE component_id = ?");
+        $resultado = $sentencia->execute([$idComp]);
+        echo json_encode($resultado);
+    } catch (PDOException $e) {
+        error_log('PDOException - ' . $e->getMessage(), 0);
+        http_response_code(500);
+        exit("Error al conectar con la base de datos");
+    }
+} else {
+    http_response_code(400);
+    exit("Error al procesar la petición");
+}
